changing units to float64 as default for most usage so I dont have to do so much conversion, also adding utils to graphics so I can start to build it out more

+++ example/main.go
@@ -16,7 +16,7 @@ var (
 	tree       *gfx.Image
 	ttf        *gfx.Font
 	image_font *gfx.Font
-	mx, my     float32
+	mx, my     float64
 )
 
 func main() {
