+++ gfx/graphics.go
@@ -14,19 +14,19 @@ var (
 	height  = 0
 )
 
-func Circle(mode string, x, y, radius float32) {
+func Circle(mode string, x, y, radius float64) {
 	Circlep(mode, x, y, radius, defaultPointCount)
 }
 
-func Circlep(mode string, x, y, radius float32, points int) {
+func Circlep(mode string, x, y, radius float64, points int) {
 	Ellipsep(mode, x, y, radius, radius, points)
 }
 
-func Arc(mode string, x, y, radius, angle1, angle2 float32) {
+func Arc(mode string, x, y, radius, angle1, angle2 float64) {
 	Arcp(mode, x, y, radius, angle1, angle2, defaultPointCount)
 }
 
-func Arcp(mode string, x, y, radius, angle1, angle2 float32, points int) {
+func Arcp(mode string, x, y, radius, angle1, angle2 float64, points int) {
 	// Nothing to display with no points or equal angles. (Or is there with line mode?)
 	if points <= 0 || angle1 == angle2 {
 		return
@@ -38,7 +38,7 @@ func Arcp(mode string, x, y, radius, angle1, angle2 float32, points int) {
 		return
 	}
 
-	angle_shift := (angle2 - angle1) / float32(points)
+	angle_shift := (angle2 - angle1) / float64(points)
 	// Bail on precision issues.
 	if angle_shift == 0.0 {
 		return
@@ -46,7 +46,7 @@ func Arcp(mode string, x, y, radius, angle1, angle2 float32, points int) {
 
 	phi := angle1
 	num_coords := (points + 3) * 2
-	coords := make([]float32, num_coords)
+	coords := make([]float64, num_coords)
 	coords[0] = x
 	coords[num_coords-2] = x
 	coords[1] = y
@@ -54,18 +54,18 @@ func Arcp(mode string, x, y, radius, angle1, angle2 float32, points int) {
 
 	for i := 0; i <= points; i++ {
 		phi = phi + angle_shift
-		coords[2*(i+1)] = x + radius*float32(math.Cos(float64(phi)))
-		coords[2*(i+1)+1] = y + radius*float32(math.Sin(float64(phi)))
+		coords[2*(i+1)] = x + radius*math.Cos(float64(phi))
+		coords[2*(i+1)+1] = y + radius*math.Sin(float64(phi))
 	}
 
 	Polygon(mode, coords)
 }
 
-func Ellipse(mode string, x, y, a, b float32) {
+func Ellipse(mode string, x, y, a, b float64) {
 	Ellipsep(mode, x, y, a, b, defaultPointCount)
 }
 
-func Ellipsep(mode string, x, y, a, b float32, points int) {
+func Ellipsep(mode string, x, y, a, b float64, points int) {
 	two_pi := math.Pi * 2.0
 	if points <= 0 {
 		points = 1
@@ -74,11 +74,11 @@ func Ellipsep(mode string, x, y, a, b float32, points int) {
 	angle_shift := two_pi / float64(points)
 	phi := 0.0
 
-	coords := make([]float32, 2*(points+1))
+	coords := make([]float64, 2*(points+1))
 	for i := 0; i < points; i++ {
 		phi += angle_shift
-		coords[2*i+0] = x + a*float32(math.Cos(float64(phi)))
-		coords[2*i+1] = y + b*float32(math.Sin(float64(phi)))
+		coords[2*i+0] = x + a*math.Cos(phi)
+		coords[2*i+1] = y + b*math.Sin(phi)
 	}
 
 	coords[2*points+0] = coords[0]
@@ -87,37 +87,37 @@ func Ellipsep(mode string, x, y, a, b float32, points int) {
 	Polygon(mode, coords)
 }
 
-func Line(args ...float32) {
+func Line(args ...float64) {
 	PolyLine(args)
 }
 
-func PolyLine(coords []float32) {
+func PolyLine(coords []float64) {
 	gl.EnableVertexAttribArray(0)
-	gl.VertexAttribPointer(0, 2, gl.FLOAT, false, 0, gl.Ptr(coords))
+	gl.VertexAttribPointer(0, 2, gl.DOUBLE, false, 0, gl.Ptr(coords))
 	gl.DrawArrays(gl.LINE_STRIP, 0, int32(len(coords))/2)
 	gl.DisableVertexAttribArray(0)
 }
 
-func Rect(mode string, x, y, w, h float32) {
-	Polygon(mode, []float32{x, y, x, y + h, x + w, y + h, x + w, y, x, y})
+func Rect(mode string, x, y, w, h float64) {
+	Polygon(mode, []float64{x, y, x, y + h, x + w, y + h, x + w, y, x, y})
 }
 
-func Polygon(mode string, coords []float32) {
+func Polygon(mode string, coords []float64) {
 	if mode == "line" {
 		PolyLine(coords)
 	} else {
 		gl.EnableVertexAttribArray(0)
-		gl.VertexAttribPointer(0, 2, gl.FLOAT, false, 0, gl.Ptr(coords))
+		gl.VertexAttribPointer(0, 2, gl.DOUBLE, false, 0, gl.Ptr(coords))
 		gl.DrawArrays(gl.TRIANGLE_FAN, 0, int32(len(coords))/2-1)
 		gl.DisableVertexAttribArray(0)
 	}
 }
 
-func DrawS(drawable Drawable, x, y float32) {
+func DrawS(drawable Drawable, x, y float64) {
 	drawable.Draw(x, y, 0, 0, 0, 0, 0, 0, 0)
 }
 
-func Draw(drawable Drawable, x, y, angle, sx, sy, ox, oy, kx, ky float32) {
+func Draw(drawable Drawable, x, y, angle, sx, sy, ox, oy, kx, ky float64) {
 	drawable.Draw(x, y, angle, sx, sy, ox, oy, kx, ky)
 }
